+++ src/main/java/restaurant/manager/RestaurantWebServiceWithDatabase/Controllers/ReservationController.java
@@ -5,11 +5,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import restaurant.manager.RestaurantWebServiceWithDatabase.DTOs.ReservationDTO;
 import restaurant.manager.RestaurantWebServiceWithDatabase.Entities.Reservation;
+import restaurant.manager.RestaurantWebServiceWithDatabase.Exceptions.ForbiddenException;
 import restaurant.manager.RestaurantWebServiceWithDatabase.Exceptions.NotFoundException;
 import restaurant.manager.RestaurantWebServiceWithDatabase.Services.ReservationService;
 import restaurant.manager.RestaurantWebServiceWithDatabase.Utilities.OkResponse;
-import restaurant.manager.RestaurantWebServiceWithDatabase.Utilities.OrderStatus;
 import restaurant.manager.RestaurantWebServiceWithDatabase.Utilities.Views;
 
 import java.util.List;
@@ -36,7 +37,11 @@ public class ReservationController {
     @GetMapping("/restaurant/{id}")
     public List<Reservation> getAllReservationsPerRestaurant(@RequestParam(defaultValue = "0") Integer page,
                                                              @RequestParam(defaultValue = "5") Integer itemsPerPage,
-                                                             @PathVariable int id) {
+                                                             @PathVariable Integer id,
+                                                             @RequestAttribute("client-username") String clientUsername) {
+        if (!reservationService.isUserTheOwnerOfRestaurant(clientUsername, id)) {
+            throw new ForbiddenException("The user is not the owner of this restaurant so he/she can not see all the reservation made for this restaurant.");
+        }
         List<Reservation> fetchedReservations = reservationService.fetchAllReservationsOfOneRestaurant(page, itemsPerPage, id);
         if (fetchedReservations.isEmpty()) {
             throw new NotFoundException("No Reservation was found");
@@ -48,7 +53,11 @@ public class ReservationController {
     @GetMapping("/user/{id}")
     public List<Reservation> getAllReservationsPerUser(@RequestParam(defaultValue = "0") Integer page,
                                                        @RequestParam(defaultValue = "5") Integer itemsPerPage,
-                                                       @PathVariable int id) {
+                                                       @PathVariable Integer id,
+                                                       @RequestAttribute("client-username") String clientUsername) {
+        if (!reservationService.isUserGivingHisOwnId(clientUsername, id)) {
+            throw new ForbiddenException("Access Denied");
+        }
         List<Reservation> userFetchedReservations = reservationService
                 .fetchAllReservationsOfOneUser(page, itemsPerPage, id);
         if (userFetchedReservations.isEmpty()) {
@@ -59,7 +68,11 @@ public class ReservationController {
 
     @JsonView(value = Views.Public.class)
     @GetMapping("/{id}")
-    public Reservation getOneReservation(@PathVariable int id){
+    public Reservation getOneReservation(@PathVariable Integer id,
+                                         @RequestAttribute("client-username") String clientUsername) {
+        if (!reservationService.isUserTheOwnerOfTheOrderedFoodRestaurant(clientUsername, id)) {
+            throw new ForbiddenException("The user is not the owner of this restaurant so he/she can not access the reservations of the restaurant.");
+        }
         Reservation fetchedReservation = reservationService.fetchOneReservation(id);
         if (fetchedReservation == null) {
             throw new NotFoundException("No Reservation was found");
@@ -68,31 +81,46 @@ public class ReservationController {
     }
 
     @PostMapping("/")
-    public ResponseEntity addReservation(@RequestParam Integer userId,
-                                         @RequestParam Integer foodId) {
-        reservationService.addReservation(userId, foodId);
+    public ResponseEntity addReservation(@RequestBody ReservationDTO reservationDTO,
+                                         @RequestAttribute("client-username") String clientUsername) {
+        if (reservationDTO.getUserId() == null ||
+                !reservationService.isUserGivingHisOwnId(clientUsername, reservationDTO.getUserId())) {
+            throw new ForbiddenException("Access Denied");
+        }
+        reservationService.addReservation(reservationDTO);
         return new ResponseEntity<>(new OkResponse("Reservation successfully added"), HttpStatus.OK);
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity updateReservation(@PathVariable int id,
-                                            @RequestParam Integer userId,
-                                            @RequestParam Integer foodId) {
-        reservationService.updateReservation(id, userId, foodId);
+    public ResponseEntity updateReservation(@PathVariable Integer id,
+                                            @RequestBody ReservationDTO reservationDTO,
+                                            @RequestAttribute("client-username") String clientUsername) {
+        if (!reservationService.isUserTheOwnerOfTheOrderedFoodRestaurant(clientUsername, id)) {
+            throw new ForbiddenException("The user is not the owner of this restaurant so he/she can not edit the reservations.");
+        }
+        reservationService.updateReservation(id, reservationDTO);
         return new ResponseEntity<>(new OkResponse("Reservation successfully updated"), HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity deleteReservation(@PathVariable int id) {
+    public ResponseEntity deleteReservation(@PathVariable Integer id,
+                                            @RequestAttribute("client-username") String clientUsername) {
+        if (!reservationService.isUserTheOwnerOfRestaurant(clientUsername, id)) {
+            throw new ForbiddenException("The user is not the owner of this restaurant so he/she can not delete any reservations from it.");
+        }
         reservationService.removeOneReservation(id);
-        return new ResponseEntity<>(new OkResponse("Reservation successfully added"), HttpStatus.OK);
+        return new ResponseEntity<>(new OkResponse("Reservation successfully removed"), HttpStatus.OK);
     }
 
     @PatchMapping("/{restaurantId}/{id}")
-    public ResponseEntity completeOrder(@PathVariable int restaurantId,
-                                        @PathVariable int id){
+    public ResponseEntity completeOrder(@PathVariable Integer restaurantId,
+                                        @PathVariable Integer id,
+                                        @RequestAttribute("client-username") String clientUsername) {
+        if (!reservationService.isUserTheOwnerOfRestaurant(clientUsername, restaurantId)
+                || !reservationService.doesReservationBelongToTheRestaurant(restaurantId, id)) {
+            throw new ForbiddenException("Access denied");
+        }
         reservationService.setStatusToCompleted(id);
         return new ResponseEntity<>(new OkResponse("Reservation successfully completed"), HttpStatus.OK);
     }
 }
-
